Comments Only Are Left!
After quite a mess of yesterday and overall slow progression, I can happily confirm that I have solved all layouting errors in the HTML side. It required a lot of backtracking and changes in said process (with the added frustration), but I freaking did it. Oh yeah, comments are going to get tonight and furthermore this project stack (I mean it this time).